+++ .gitignore
@@ -1,3 +1,2 @@
 *.pyc
 GameServer/.idea/
-ProxyServer/.idea/
