+++ README.md
@@ -1,5 +1,3 @@
 # CMPS183 KittyWar Server
 
-### Repo consists of two projects
 * Game Server (Pure Python)
-* Proxy Server (Pure Python)
